BAEL-2715: Modified customizers
