--- spring-5-reactive-netty/src/main/java/com/baeldung/serverconfig/server/NettyWebServerFactoryPortCustomizer.java
@@ -1,30 +1,0 @@
-package com.baeldung.serverconfig.server;
-
-import org.springframework.boot.web.embedded.netty.NettyReactiveWebServerFactory;
-import org.springframework.boot.web.embedded.netty.NettyServerCustomizer;
-import org.springframework.boot.web.server.WebServerFactoryCustomizer;
-import org.springframework.stereotype.Component;
-import reactor.netty.http.server.HttpServer;
-
-@Component
-public class NettyWebServerFactoryPortCustomizer implements WebServerFactoryCustomizer<NettyReactiveWebServerFactory> {
-
-    @Override
-    public void customize(NettyReactiveWebServerFactory serverFactory) {
-        serverFactory.addServerCustomizers(new PortCustomizer(8443));
-    }
-
-    private static class PortCustomizer implements NettyServerCustomizer {
-
-        private final int port;
-
-        private PortCustomizer(int port) {
-            this.port = port;
-        }
-
-        @Override
-        public HttpServer apply(HttpServer httpServer) {
-            return httpServer.port(port);
-        }
-    }
-}
