+++ spring-5-reactive-netty/src/main/java/com/baeldung/serverconfig/server/NettyWebServerFactorySslCustomizer.java
@@ -1,6 +1,8 @@
 package com.baeldung.serverconfig.server;
 
 import org.springframework.boot.web.embedded.netty.NettyReactiveWebServerFactory;
+import org.springframework.boot.web.embedded.netty.SslServerCustomizer;
+import org.springframework.boot.web.server.Http2;
 import org.springframework.boot.web.server.Ssl;
 import org.springframework.boot.web.server.WebServerFactoryCustomizer;
 import org.springframework.stereotype.Component;
@@ -16,7 +18,9 @@ public class NettyWebServerFactorySslCustomizer implements WebServerFactoryCusto
         ssl.setKeyAlias("alias");
         ssl.setKeyPassword("password");
         ssl.setKeyStorePassword("secret");
-        serverFactory.setSsl(ssl);
+        Http2 http2 = new Http2();
+        http2.setEnabled(false);
+        serverFactory.addServerCustomizers(new SslServerCustomizer(ssl, http2, null));
         serverFactory.setPort(8443);
     }
 }
